+++ R/scan.R
@@ -13,10 +13,10 @@
 #'
 #' @return a `scan` object (S3 class) containing:
 #' \itemize{
-#'   \item{`raw.scan`: a list of raw binary adjacency matrix shaped like the
+#'   \item{`raw.scan.list`: a list of raw binary adjacency matrix shaped like the
 #'   `Adj` contained in `presence.prob`, considered directed in the
 #'   algorithm}
-#'   \item{`theoretical.scan`: a list of binary adjacency matrix, where all ties were
+#'   \item{`theoretical.scan.list`: a list of binary adjacency matrix, where all ties were
 #'   observed but the `mode` has been applied}
 #'   \item{`scan.type`: character scalar. `generate_scan` sets it to
 #'   "theoretical", `sample_from_scan` will set it to "empirical" and append the
@@ -25,8 +25,6 @@
 #'   \item{`total_scan`: `total_scan` data contained in `presence.prob`}
 #'   \item{`scans.to.do`: inputted `scans.to.do`}
 #'   \item{`mode`: `mode` data contained in `presence.prob`}
-#'   \item{`weighted`: logical, at this stage can only be `TRUE` if `mode =
-#'   plus` (some edges can become `2`)}
 #'   \item{`Adj.subfun`: `Adj.subfun` data contained in `presence.prob`}
 #'   \item{`presence.prob`: `presence.prob$P` (only the probability matrix) data
 #'   contained in `presence.prob`}
@@ -35,16 +33,16 @@
 #' @noRd
 generate_scan<- function(presence.prob,scans.to.do){
   if(length(scans.to.do) == 1) {if(scans.to.do == "all") {scans.to.do <- 1:presence.prob$total_scan}}
-  raw.scan<- draw_raw.scan(presence.prob = presence.prob,scans.to.do = scans.to.do)
+  raw.scan.list <- draw_raw.scan.list(presence.prob = presence.prob,scans.to.do = scans.to.do)
+  theoretical.scan.list <- apply_mode(raw.scan.list = raw.scan.list,mode = presence.prob$mode)
   scan<- list(
-    raw.scan = raw.scan,
-    theoretical.scan = apply_mode(raw.scan = raw.scan,mode = presence.prob$mode),
+    raw.scan.list = raw.scan.list,
+    theoretical.scan.list = theoretical.scan.list,
     scan.type = "theoretical",
     Adj = presence.prob$Adj,
     total_scan = presence.prob$total_scan,
     scans.to.do = scans.to.do,
     mode = presence.prob$mode,
-    weighted = ifelse(presence.prob$mode == "plus",FALSE,TRUE), # only case for which an edge can be > 1
     Adj.subfun = presence.prob$Adj.subfun,
     presence.prob = presence.prob$P # in here it is not a presenceProb object anymore, to avoid storing redundant variables
   )
@@ -56,18 +54,70 @@ generate_scan<- function(presence.prob,scans.to.do){
 #' @export
 #' @noRd
 print.scan<- function(x,...){
-  cat("Scan type: ",x$scan.type,", mode: ",x$mode,"\n\n",sep = "")
-  n <- length(x$theoretical.scan)
-  if(length(x$theoretical.scan) >= 10) {
-    print.default(x$theoretical.scan[1:3],...)
+  scans.to.do <- explicit_scan.to.do(x)
+  n <- length(scans.to.do)
+  # print the general simulation info
+  if (n >= 15) {
+    scans.to.do <- paste(do.call(paste,as.list(scans.to.do)[1:5]),
+                         "...",
+                         do.call(paste,as.list(scans.to.do)[(n-4):n])
+    )
+  } else {
+    scans.to.do <- scans.to.do
+  }
+  cat("\nScan(s) performed: ",scans.to.do,sep=" ")
+  cat("\nScan type: ",x$scan.type,", mode: ",x$mode,"\n\n",sep = "")
+  n <- length(x$theoretical.scan.list)
+  if (n >= 10) {
+    print.default(x$theoretical.scan.list[1:3],...)
     cat("... (",n-5," more scans)\n\n\n",sep = "")
     cat("[[",n-1,"]]\n",sep = "")
-    print.default(x$theoretical.scan[[(n-1)]],...)
+    print.default(x$theoretical.scan.list[[(n-1)]],...)
     cat("[[",n,"]]\n",sep = "")
-    print.default(x$theoretical.scan[[n]],...)
+    print.default(x$theoretical.scan.list[[n]],...)
   } else {
-    print.default(x$theoretical.scan,...)
+    print.default(x$theoretical.scan.list,...)
   }
+}
+
+#' Summary method for `scan` objects
+#' @export
+#' @noRd
+summary.scan <- function(object,...) {
+  scans.to.do <- explicit_scan.to.do(object)
+  mode <- object$mode
+  theoretical.sum <- sum_scan.list(object$theoretical.scan.list)
+  theoretical.sampled <- sum_scan.sampled(object,method = "theoretical")
+
+  scan.summary <- list(
+    theoretical.sum = theoretical.sum,
+    theoretical.sampled = theoretical.sampled,
+    scans.to.do = scans.to.do,
+    mode = mode
+    # more things can be added here# more things can be added here
+  )
+  class(scan.summary) <- "summary.scan"
+  scan.summary
+  # more info to be displayed
+}
+
+#' Print method for `summary.scan` objects
+#' @export
+#' @noRd
+print.summary.scan<- function(x,...){
+  cat("Theoretical weighted adjacency matrix:\n")
+  print.default(x$theoretical.sum,...)
+  cat(paste0("\nobtained after summing ", length(x$scans.to.do), " binary scans (mode = \"", x$mode,"\")", "\n\n"))
+}
+
+#' Plot method for `scan` objects
+#' @importFrom igraph graph_from_adjacency_matrix
+#' @importFrom igraph plot.igraph
+#' @export
+#' @noRd
+plot.scan<- function(x,...){
+  x<- igraph::graph_from_adjacency_matrix(x$theoretical,mode = x$mode,weighted = x$weighted)
+  igraph::plot.igraph(x,...,main = "theoretical")
 }
 
 #' Plot method for `scan` objects
@@ -93,10 +143,10 @@ plot.scan<- function(x,...){
 #' @return a list of raw binary adjacency matrix shaped like the `Adj` contained in `presence.prob`
 #'
 #' @noRd
-draw_raw.scan <- function(presence.prob,scans.to.do){
+draw_raw.scan.list <- function(presence.prob,scans.to.do){
   n<- nrow(presence.prob$Adj);nodes_names<- rownames(presence.prob$Adj)
   presence.P.vec<- presence.prob$P[presence.prob$Adj.subfun(presence.prob$P)]; p<- length(presence.P.vec)  # subset a presence probability vector the same way Adj is subset (depends on `Adj`'s mode (cf. igraph)) <- NOW THIS IS JUST A UPPER/LOWER.TRI VS NON.DIAGONAL THING...
-  raw.scan <-
+  raw.scan.list <-
     rep(
       list( # required for rep to output a list
         matrix(0,nrow = n,ncol = n,dimnames = list(nodes_names,nodes_names))  # structure the scan as a matrix filled with zeros
@@ -104,9 +154,9 @@ draw_raw.scan <- function(presence.prob,scans.to.do){
       length(scans.to.do)
     )
   lapply(
-    raw.scan,
+    raw.scan.list,
     function(s) {
-      s[presence.prob$Adj.subfun(s)]<- stats::rbinom(p,1,presence.P.vec)  # core of the randomization: draw a (raw.scan) tie or not for each (relevant, cf. triangular matrices or undirected) dyad according to its presence probability
+      s[presence.prob$Adj.subfun(s)]<- stats::rbinom(p,1,presence.P.vec)  # core of the randomization: draw a (raw.scan.list) tie or not for each (relevant, cf. triangular matrices or undirected) dyad according to its presence probability
       s
     }
   )